+++ CMakeProject1/file_func.cpp
@@ -14,16 +14,16 @@ int count_lines(std::string fileName) {
 	return i;
 }
 
-std::vector<int> copy_data(std::string fileName, int start, int limit) {
+std::vector<float> copy_data(std::string fileName, int start, int limit) {
 	std::cout << "starting copy" << std::endl;
-	std::vector<int> num_arry(limit);
+	std::vector<float> num_arry(limit);
 	std::string my_text;
 	std::ifstream MyReadFile(fileName);
 	int i = 0;
 	while (std::getline(MyReadFile, my_text)) {
 		if ((i < limit) && (i >= start)) {
 			try {
-				num_arry[i] = std::stoi(my_text);
+				num_arry[i] = std::stof(my_text);
 			}
 			catch (const std::invalid_argument& e) {
 				std::cout << "Invalid input: " << e.what() << std::endl;
@@ -39,7 +39,7 @@ std::vector<int> copy_data(std::string fileName, int start, int limit) {
 	return num_arry;
 }
 
-int write_result(std::vector<int> data, int start, int end) {
+int write_result(std::vector<float> data, int start, int end) {
 	std::cout << "writing data..." << std::endl;
 	int result = 0;
 	std::string fileName = "results.txt";
